+++ api/api/create_database.py
@@ -24,7 +24,6 @@ if __name__ == "__main__":
     ## Create the new tables and populate with initial data
     query_files = list(PUZZLE_CREATE_TABLE_LIST)
     query_files.append("initial_puzzle_variant.sql")
-    query_files.append("insert_initial_bit_expiration_round_2.sql")
     query_files.append("insert_initial_anon_user.sql")
 
     for file_path in query_files: